api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "fastapi[standard]>=0.116.1",
    "google-cloud-storage>=3.2.0",
    "gunicorn>=23.0.0",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
